requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.9.3
    # via
    #   eth-brownie
    #   web3
aiosignal==1.3.1
    # via
    #   aiohttp
    #   eth-brownie
asttokens==2.4.1
    # via
    #   eth-brownie
    #   vyper
attrs==23.2.0
    # via
    #   aiohttp
    #   eth-brownie
    #   hypothesis
    #   jsonschema
    #   pytest
    #   referencing
bitarray==2.9.2
    # via
    #   eth-account
    #   eth-brownie
black==24.2.0
    # via
    #   -r requirements.in
    #   eth-brownie
brownie-token-tester==0.3.2
    # via -r requirements.in
build==1.2.1
    # via pip-tools
cbor2==5.6.2
    # via
    #   eth-brownie
    #   vyper
certifi==2024.2.2
    # via
    #   eth-brownie
    #   requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via
    #   eth-brownie
    #   requests
click==8.1.7
    # via
    #   black
    #   eth-brownie
    #   pip-tools
cytoolz==0.12.3
    # via
    #   eth-brownie
    #   eth-utils
dataclassy==0.11.1
    # via
    #   eip712
    #   eth-brownie
distlib==0.3.8
    # via virtualenv
eip712==0.2.4
    # via eth-brownie
eth-abi==5.0.0
    # via
    #   eip712
    #   eth-account
    #   eth-brownie
    #   eth-event
    #   web3
eth-account==0.10.0
    # via
    #   eip712
    #   eth-brownie
    #   web3
eth-brownie==1.20.3
    # via
    #   -r requirements.in
    #   brownie-token-tester
eth-event==1.2.5
    # via eth-brownie
eth-hash[pycryptodome]==0.6.0
    # via
    #   eip712
    #   eth-brownie
    #   eth-event
    #   eth-utils
    #   web3
eth-keyfile==0.7.0
    # via
    #   eth-account
    #   eth-brownie
eth-keys==0.5.0
    # via
    #   eth-account
    #   eth-brownie
    #   eth-keyfile
eth-rlp==1.0.1
    # via
    #   eth-account
    #   eth-brownie
eth-typing==3.5.2
    # via
    #   eip712
    #   eth-abi
    #   eth-brownie
    #   eth-keys
    #   eth-utils
    #   web3
eth-utils==2.3.1
    # via
    #   eip712
    #   eth-abi
    #   eth-account
    #   eth-brownie
    #   eth-event
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   rlp
    #   web3
execnet==2.0.2
    # via
    #   eth-brownie
    #   pytest-xdist
filelock==3.14.0
    # via virtualenv
flake8==3.9.2
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
    #   eth-brownie
hexbytes==0.3.1
    # via
    #   eip712
    #   eth-account
    #   eth-brownie
    #   eth-event
    #   eth-rlp
    #   web3
hypothesis==6.27.3
    # via eth-brownie
identify==2.5.36
    # via pre-commit
idna==3.6
    # via
    #   eth-brownie
    #   requests
    #   yarl
importlib-metadata==7.0.1
    # via
    #   eth-brownie
    #   vyper
iniconfig==2.0.0
    # via
    #   eth-brownie
    #   pytest
isort==5.13.2
    # via -r requirements.in
jsonschema==4.21.1
    # via
    #   eth-brownie
    #   web3
jsonschema-specifications==2023.12.1
    # via
    #   eth-brownie
    #   jsonschema
lazy-object-proxy==1.10.0
    # via eth-brownie
lru-dict==1.2.0
    # via
    #   eth-brownie
    #   web3
mccabe==0.6.1
    # via flake8
multidict==6.0.5
    # via
    #   aiohttp
    #   eth-brownie
    #   yarl
mypy-extensions==1.0.0
    # via
    #   black
    #   eth-brownie
nodeenv==1.9.0
    # via pre-commit
packaging==23.2
    # via
    #   black
    #   build
    #   eth-brownie
    #   pytest
    #   vyper
parsimonious==0.9.0
    # via
    #   eth-abi
    #   eth-brownie
pathspec==0.12.1
    # via
    #   black
    #   eth-brownie
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.2.0
    # via
    #   black
    #   eth-brownie
    #   virtualenv
pluggy==1.4.0
    # via
    #   eth-brownie
    #   pytest
pre-commit==3.7.1
    # via -r requirements.in
prompt-toolkit==3.0.43
    # via eth-brownie
protobuf==4.25.3
    # via
    #   eth-brownie
    #   web3
psutil==5.9.8
    # via eth-brownie
py==1.11.0
    # via
    #   eth-brownie
    #   pytest
    #   pytest-forked
py-solc-ast==1.2.10
    # via eth-brownie
py-solc-x==1.1.1
    # via eth-brownie
pycodestyle==2.7.0
    # via flake8
pycryptodome==3.20.0
    # via
    #   eth-brownie
    #   eth-hash
    #   eth-keyfile
    #   vyper
pyflakes==2.3.1
    # via flake8
pygments==2.17.2
    # via
    #   eth-brownie
    #   pygments-lexer-solidity
pygments-lexer-solidity==0.7.0
    # via eth-brownie
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pytest==6.2.5
    # via
    #   eth-brownie
    #   pytest-forked
    #   pytest-xdist
pytest-forked==1.6.0
    # via
    #   eth-brownie
    #   pytest-xdist
pytest-xdist==1.34.0
    # via eth-brownie
python-dotenv==0.16.0
    # via eth-brownie
pyunormalize==15.1.0
    # via
    #   eth-brownie
    #   web3
pyyaml==6.0.1
    # via
    #   eth-brownie
    #   pre-commit
referencing==0.33.0
    # via
    #   eth-brownie
    #   jsonschema
    #   jsonschema-specifications
regex==2023.12.25
    # via
    #   eth-brownie
    #   parsimonious
requests==2.31.0
    # via
    #   eth-brownie
    #   py-solc-x
    #   vvm
    #   web3
rlp==4.0.0
    # via
    #   eth-account
    #   eth-brownie
    #   eth-rlp
rpds-py==0.18.0
    # via
    #   eth-brownie
    #   jsonschema
    #   referencing
semantic-version==2.10.0
    # via
    #   eth-brownie
    #   py-solc-x
    #   vvm
six==1.16.0
    # via
    #   asttokens
    #   eth-brownie
    #   pytest-xdist
sortedcontainers==2.4.0
    # via
    #   eth-brownie
    #   hypothesis
toml==0.10.2
    # via
    #   eth-brownie
    #   pytest
toolz==0.12.1
    # via
    #   cytoolz
    #   eth-brownie
tqdm==4.66.2
    # via eth-brownie
typing-extensions==4.9.0
    # via
    #   eth-brownie
    #   eth-rlp
    #   eth-typing
    #   web3
urllib3==2.2.1
    # via
    #   eth-brownie
    #   requests
virtualenv==20.26.2
    # via pre-commit
vvm==0.1.0
    # via eth-brownie
vyper==0.3.10
    # via eth-brownie
wcwidth==0.2.13
    # via
    #   eth-brownie
    #   prompt-toolkit
web3==6.15.1
    # via eth-brownie
websockets==12.0
    # via
    #   eth-brownie
    #   web3
wheel==0.42.0
    # via
    #   eth-brownie
    #   pip-tools
    #   vyper
wrapt==1.16.0
    # via eth-brownie
yarl==1.9.4
    # via
    #   aiohttp
    #   eth-brownie
zipp==3.17.0
    # via
    #   eth-brownie
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
